Update shell.bat
Made it cd to your home directory by default and made a "app" command that launches apps on your computer and in the folder that the shell.bat file is in (like the program "chrome" or if you have a file that is in the same folder as shell.bat called "image.png" that will also work). This also supports links to files Also made a command that changes the directory into the folder that the shell.bat is in.