Theme Crash Recovery
Made it to where batShell can recover from a theme crash. If batShell suddenly forgets if it has a theme or not, it will pause the cmd prompt and ask if the user wants to turn off the theme for that session. If the user chooses to disable themes, they can keep working without worrying on the theme ruining their productive work.